requirements.txt
streamlit
pypdf
python-docx
Pillow
pytesseract
openpyxl